refactor(oracle): move tests to integration tests and add timing guard test
